* Moved forms (source files) into 'Forms' folders in all wizards. * Added display of version number '2019.0' in dialog title of Code Generation Wizard and Upgrade Wizards. * Removed some unused namespaces. * Removed some extraneous white space in code files. * Filled in some "TODO" entries in method headers. * Added back post build event handlers for UIWizard and UpgradeWizard to copy final vsix to \WebSDK\bin\Wizards\ folder. * Solution Wizard - Added read-only display of an example of what a namespace will look like after specifying Company Name, ModuleID (also called AppId) and Namespace (if overridden).
