+++ src/wizards/CodeGenerationWizard/EnumValue.cs
@@ -69,6 +69,5 @@ namespace Sage.CA.SBS.ERP.Sage300.CodeGenerationWizard
             return string.Empty;
         }
         #endregion
-
     }
 }
