+++ src/wizards/Sage300SolutionWizard/Properties/Resources.resx
@@ -248,6 +248,6 @@ prior to deployment.</value>
     <value>Spanish</value>
   </data>
   <data name="SolutionGeneration" xml:space="preserve">
-    <value>Solution Generation</value>
+    <value>Solution Generation Wizard</value>
   </data>
 </root>
