+++ src/wizards/Sage300UpgradeWizard/Sage300UpgradeWizardPackage/Sage300UpgradeWizardPackage.csproj
@@ -111,8 +111,7 @@
 REM Call RebuildPayload.bat</PreBuildEvent>
   </PropertyGroup>
   <PropertyGroup>
-    <PostBuildEvent>
-    </PostBuildEvent>
+    <PostBuildEvent>copy $(TargetDir)$(TargetName).vsix $(ProjectDir)..\..\..\..\bin\wizards\</PostBuildEvent>
   </PropertyGroup>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
